pyproject.toml
[tool.poetry]
name = "gridmaster-scenario-generation"
version = "0.1.0"
description = ""

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.4"
PyMySQL = "^1.0.2"
openpyxl = "^3.0.9"
pyESDL = "^21.11.0"
s3fs = "^2021.11.1"
boto3 = ">=1.19.0"
pyecore = "0.12.1"
pyecoregen = "^0.5.1"
ema-workbench = "^2.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
